+++ examples/cpp/broker.hpp
@@ -151,7 +151,7 @@ class queues {
   protected:
     typedef std::map<std::string, queue *> queue_map;
     queue_map queues_;
-    uint64_t next_id_; // Use to generate unique queue IDs.
+    int next_id_; // Use to generate unique queue IDs.
 };
 
 #include <proton/config.hpp>
